Properly setup github changelog generator
We used it via command line previously for 1 off generation. Instead lets add it to the Rakefile with all the options. I went through and tagged EVERY pull request back to the beginning. This gets us enhancements/bugs and not just “merged PR” since a lot of them were total tech cleanup that shouldn’t be in the changelog.

Signed-off-by: Tim Smith <tsmith@chef.io>